Add file io redirection
Redefined parse_tokens to parse_command to manage more complex
commands involving io redirection with '<' or '>' symbols.

Created one structure command consisting of the tokens and the
io INPUT AND OUTPUT of the command.

Better to have it together in a struct for each command. Later if we
add functionality for pipelines, it will be easier to handle a list
of command_types.

All i/o routines have been put in a io.c file.